\name{diyabcGUI-package}
\alias{diyabcGUI-package}
\alias{diyabcGUI}
\docType{package}
\title{
\packageTitle{diyabcGUI}
}
\description{
\packageDescription{diyabcGUI}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{diyabcGUI}
\packageIndices{diyabcGUI}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{diyabcGUI}

Maintainer: \packageMaintainer{diyabcGUI}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
